requirements.txt
Django==1.11.1
pygments==2.2.0
django-el-pagination==3.1.0
django-mathfilters==0.4.0